+++ app/controllers/registrations_controller.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-# Updates layout for devise registrations pages.
+# Override for devise registrations controller.
 class RegistrationsController < Devise::RegistrationsController
   layout "layouts/full_page"
 end