+++ app/models/pg_search_document.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+# Add methods to existing PgSearch::Document model.
+class PgSearchDocument < PgSearch::Document
+  include Searchable
+
+  def self.searchable_attributes
+    %w(content)
+  end
+end