+++ config/routes.rb
@@ -22,6 +22,8 @@ Rails.application.routes.draw do
 
   resources :drugs
 
+  get "documents" => "documents#index"
+
   namespace :editor do
     resources :categories, path: "folders" do
       resources :documents do
@@ -120,6 +122,7 @@ Rails.application.routes.draw do
     get :dashboard
     get :directory
     get :leaving
+    get :search
     get ":top_level/:category", action: :category, as: :internal_category
     get ":top_level/:category/documents/:document_id", action: :document, as: :internal_category_document
     get ":top_level/:category/videos/:video_id", action: :video, as: :internal_category_video