+++ gems.rb
@@ -18,6 +18,7 @@ gem "font-awesome-sass",          "~> 5.6.1"
 gem "haml",                       "~> 5.0.4"
 gem "kaminari",                   "~> 1.1.1"
 gem "mini_magick",                "~> 4.9.2"
+gem "pg_search",                  "~> 2.1.4"
 
 # Rails defaults.
 gem "coffee-rails",               "~> 4.2"