+++ test/controllers/external_controller_test.rb
@@ -6,16 +6,9 @@ require "test_helper"
 class ExternalControllerTest < ActionDispatch::IntegrationTest
   setup do
     @viewer = users(:viewer)
-    @generic_viewer = viewers(:one)
   end
 
   test "should get contact as public viewer" do
-    get contact_url
-    assert_response :success
-  end
-
-  test "should get contact as generic viewer" do
-    login_viewer(@generic_viewer)
     get contact_url
     assert_response :success
   end
@@ -31,12 +24,6 @@ class ExternalControllerTest < ActionDispatch::IntegrationTest
     assert_response :success
   end
 
-  test "should get landing as generic viewer" do
-    login_viewer(@generic_viewer)
-    get landing_url
-    assert_response :success
-  end
-
   test "should get landing as viewer" do
     login(@viewer)
     get landing_url
@@ -48,12 +35,6 @@ class ExternalControllerTest < ActionDispatch::IntegrationTest
     assert_response :success
   end
 
-  test "should get participate as generic viewer" do
-    login_viewer(@generic_viewer)
-    get participate_url
-    assert_response :success
-  end
-
   test "should get participate as viewer" do
     login(@viewer)
     get participate_url
@@ -61,12 +42,6 @@ class ExternalControllerTest < ActionDispatch::IntegrationTest
   end
 
   test "should get sites as public viewer" do
-    get sites_url
-    assert_response :success
-  end
-
-  test "should get sites as generic viewer" do
-    login_viewer(@generic_viewer)
     get sites_url
     assert_response :success
   end