+++ test/controllers/setup/setup_controller_test.rb
@@ -8,7 +8,6 @@ class Setup::SetupControllerTest < ActionDispatch::IntegrationTest
     @admin = users(:admin)
     @editor = users(:editor)
     @viewer = users(:viewer)
-    @generic_viewer = viewers(:one)
   end
 
   test "should get index as admin" do
@@ -27,12 +26,6 @@ class Setup::SetupControllerTest < ActionDispatch::IntegrationTest
     login(@viewer)
     get setup_url
     assert_redirected_to dashboard_url
-  end
-
-  test "should not get index as generic viewer" do
-    login_viewer(@generic_viewer)
-    get setup_url
-    assert_redirected_to new_user_session_url
   end
 
   test "should not get index as public viewer" do