+++ test/test_helper.rb
@@ -25,13 +25,9 @@ class ActionController::TestCase
     @request.env["devise.mapping"] = Devise.mappings[resource]
     sign_in(resource, scope: resource.class.name.downcase.to_sym)
   end
-
-  def login_viewer(resource)
-    session[:viewer_id] = resource.id
-  end
 end
 
-# Set up ActionDispatch tests
+# Set up ActionDispatch tests.
 class ActionDispatch::IntegrationTest
   include Devise::Test::IntegrationHelpers
 
@@ -39,22 +35,13 @@ class ActionDispatch::IntegrationTest
     sign_in_as(user, "1234567890")
   end
 
-  def login_viewer(viewer)
-    sign_in_as_viewer(viewer)
-  end
-
   def sign_in_as(user, password)
     user.update password: password, password_confirmation: password
-    post new_user_session_url, params: { user: { email: user.email, password: password } }
+    post new_user_session_url, params: {
+      user: { email: user.email, password: password }
+    }
     follow_redirect!
     user
-  end
-
-  def sign_in_as_viewer(viewer)
-    viewer.update password: viewer.password_plain, password_confirmation: viewer.password_plain
-    post new_user_session_url, params: { user: { email: viewer.username, password: viewer.password_plain } }
-    follow_redirect!
-    viewer
   end
 end